actually use the binder `replicas` value
`replicas` is a binder value in the `values.yaml` but the service just uses `replicas: 1`.  I think its intentional that you can configure replicas as you might want multiple in large cluster.

This PR updates the `binder.yaml` to use the `replicas` value for `binder`.

Should be a no-op because the default value is `1` but allows configuration if needed.